image_noise_project/requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==5.1.1
djongo==1.2.31
dnspython==2.6.1
pillow==10.4.0
pymongo==4.9.1
sqlparse==0.5.1
tzdata==2024.2
